add read/write timeout and change timeout to double

+++ hs_response.h
@@ -10,7 +10,7 @@
 #define HS_CODE_ESCAPE_PREFIX   0x01
 #define HS_CODE_ESCAPE_ADD      0x40
 
-int hs_response_value(php_stream *stream, long timeout, zval *return_value, zval *error, int modify TSRMLS_DC);
-int hs_response_multi(php_stream *stream, long timeout, zval *return_value, zval *error, zval *mreq TSRMLS_DC);
+int hs_response_value(php_stream *stream, double timeout, zval *return_value, zval *error, int modify TSRMLS_DC);
+int hs_response_multi(php_stream *stream, double timeout, zval *return_value, zval *error, zval *mreq TSRMLS_DC);
 
 #endif /* HANDLERSOCKET_RESPONSE_H */
